requirements.txt
aniso8601==3.0.2
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.5.0
Django==1.11.16
django-cors-headers==2.4.0
django-filter==2.1.0
django-graphene-permissions==0.0.4
django-graphql-jwt==0.2.1
djangorestframework==3.9.0
enum34==1.1.6
graphene==2.1.3
graphene-django==2.2.0
graphene-mongo==0.1.19
graphene-permissions==1.1.2
graphql-core==2.1
graphql-relay==0.4.5
iso8601==0.1.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mongoengine==0.17.0
promise==2.2.1
protobuf==3.6.1
psycopg2==2.8.2
PyJWT==1.7.1
pylint==1.9.3
pymongo==3.7.2
pytz==2018.7
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.6
wrapt==1.10.11
